Refactor the TypeResolution pass to use fine grained visitors
This is a huge refactor but no logic has been changed. Most implementation
has now been moved from the headers and into cc files now. This should
help alot with compilation times when working within the type system.

This is another incremental step to improve our coercion site logic as also
support forward declared items.

Addresses #1455
