+++ gcc/rust/typecheck/rust-hir-type-check-struct-field.h
@@ -29,31 +29,18 @@ namespace Resolver {
 
 class TypeCheckStructExpr : public TypeCheckBase
 {
-  using Rust::Resolver::TypeCheckBase::visit;
-
 public:
-  static TyTy::BaseType *Resolve (HIR::StructExprStructFields *expr)
-  {
-    TypeCheckStructExpr resolver (expr);
-    expr->accept_vis (resolver);
-    return resolver.resolved;
-  }
+  static TyTy::BaseType *Resolve (HIR::StructExprStructFields *expr);
 
-  void visit (HIR::StructExprStructFields &struct_expr) override;
+protected:
+  void resolve (HIR::StructExprStructFields &struct_expr);
 
-  void visit (HIR::StructExprFieldIdentifierValue &field) override;
-
-  void visit (HIR::StructExprFieldIndexValue &field) override;
-
-  void visit (HIR::StructExprFieldIdentifier &field) override;
+  void visit (HIR::StructExprFieldIdentifierValue &field);
+  void visit (HIR::StructExprFieldIndexValue &field);
+  void visit (HIR::StructExprFieldIdentifier &field);
 
 private:
-  TypeCheckStructExpr (HIR::Expr *e)
-    : TypeCheckBase (),
-      resolved (new TyTy::ErrorType (e->get_mappings ().get_hirid ())),
-      struct_path_resolved (nullptr),
-      variant (&TyTy::VariantDef::get_error_node ())
-  {}
+  TypeCheckStructExpr (HIR::Expr *e);
 
   // result
   TyTy::BaseType *resolved;
